Address comments from shuhei, remove hardcoding and fix bug in dummy eval function
